Consolidate Python interfaces into an installable package
