+++ .gitignore
@@ -29,6 +29,7 @@
 *.app
 __pycache__/*
 *.pyc
+*.egg-info
 
 # Build specific files
 build*/
